add real market data test
