+++ examples/async_all_candles.py
@@ -12,6 +12,7 @@ async def main():
         async for candle in client.get_all_candles(
             figi="BBG004730N88",
             from_=datetime.utcnow() - timedelta(days=365),
+            to=datetime.utcnow() - timedelta(days=364),
             interval=CandleInterval.CANDLE_INTERVAL_HOUR,
         ):
             print(candle)
